+++ ar-translator/UIImage+rotate.swift
@@ -8,7 +8,7 @@
 import UIKit
 
 extension UIImage {
-    func rotate(radians: Float) -> UIImage? {
+    func rotate(radians: Double) -> UIImage? {
         var newSize = CGRect(origin: CGPoint.zero, size: self.size).applying(CGAffineTransform(rotationAngle: CGFloat(radians))).size
         // Trim off the extremely small float value to prevent core graphics from rounding it up
         newSize.width = floor(newSize.width)
